Major Overhaul of the program
The biggest change was a change to the location tracking system which
had to be abstracted so that all operations are hidden behind macros.
With the new additions, there has been a spike in memory consumption.

A new, advanced test has been added that computes prime numbers.  If
something were to go wrong, then this test is the most likely to
fail.

There is still the issue of register allocation in which sometimes
after a register has been allocated, it will never be freed.  This is
must be corrected in the near future.