Altered the testing environment.
Separated the existing tests into a sub-directory of the tests
directory called `working-programs'.  This way, all the working
programs can be controlled by a single testing framework.  Also, the
test files are now only C source files (not shell scripts) so there is
no more need for any of the boiler plate that filled up the previous
prog-* files.  The new tester.sh script now takes a single argument,
the name of the source file, and runs the test on it.