prog-15: new feature where conditionals can return ints.
The code for translating a conditional of the form `a < b' was moved
to the `gen_code_binary' routine, here a test was added to an
additional field in the AST structure (`actlikejump') that if a return
value was needed then it would emit a conditional move instruction.
Otherwise, a calling `gen_code_cond' would emit a conditional jump.