bootstrap.conf: altered resulting Makefile.in.in
Gettext was complaining about version mismatch between the macros in
m4/po.m4 and the variable settings in Makefile.in.in.  As such, a sed
command was added to the bootstrap_post_import_hook to fix the setting
found in Makefile.in.in to the one that matches m4/po.m4.